refactor: Align module structure with api
