+++ dsp-meta/src/api/handler/health.rs
@@ -1,6 +1,6 @@
 use tracing::trace;
 
-pub(crate) async fn health_handler() -> &'static str {
+pub(crate) async fn health() -> &'static str {
     trace!("entered health_handler()");
     "healthy"
 }
