+++ dsp-meta/src/api/handler/mod.rs
@@ -1,4 +1,4 @@
 pub mod health;
-pub mod project_metadata_handler;
-pub mod sitemap_xml_handler;
-pub mod robots_txt_handler;
+pub mod v1;
+pub mod sitemap_xml;
+pub mod robots_txt;
