+++ dsp-meta/src/api/mod.rs
@@ -1,4 +1,3 @@
 pub mod convert;
 mod handler;
-mod model;
 pub mod router;
