--- dsp-meta/src/api/model/mod.rs
@@ -1,2 +1,0 @@
-pub mod project_metadata_dto;
-pub mod project_metadata_graph_dto;
